requirements.txt
absl-py==1.2.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
docker-pycreds==0.4.0
easyocr==1.5.0
entrypoints==0.4
executing==0.9.0
filelock==3.7.1
fonttools==4.34.4
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.21.1
ipykernel==6.9.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.9.0
setproctitle==1.3.0
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
tifffile==2022.8.12
torch==1.11.0
torch-tb-profiler==0.4.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.12.21
wcwidth==0.2.5
Werkzeug==2.2.0
wheel==0.37.1
